improve report

+++ R/util.r
@@ -806,7 +806,7 @@ ezCbind = function(...){
   do.call(cbind, x)
 }
 
-makeRmdReport = function(..., htmlFile="00index.html", rmdFile='', linkLibs=TRUE){
+makeRmdReport = function(..., htmlFile="00index.html", rmdFile='', linkHtmlLibDir=NULL){
   varList = list(...)
   for (nm in names(varList)){
     saveRDS(varList[[nm]], file=paste0(nm, ".rds"))
@@ -818,17 +818,17 @@ makeRmdReport = function(..., htmlFile="00index.html", rmdFile='', linkLibs=TRUE
   file.copy(from=styleFiles, to=".", overwrite=TRUE)
   rmarkdown::render(input=rmdFile, envir = new.env(),
                     output_dir=".", output_file=htmlFile, quiet=TRUE)
-  if (linkLibs){
-    prepareRmdLib()
-  }
+  prepareRmdLib(linkHtmlLibDir=linkHtmlLibDir)
 }
 
-prepareRmdLib <- function(){
+prepareRmdLib <- function(linkHtmlLibDir=NULL){
   ## Link the rmarkdownLib
-  file.copy(from=list.files("rmarkdownLib", full.names = TRUE),
-            to="/srv/GT/reference/rmarkdownLib",
-            recursive=TRUE, overwrite=FALSE)
-  unlink("rmarkdownLib", recursive = TRUE)
-  file.symlink(from = "/srv/GT/reference/rmarkdownLib",
-               to = "rmarkdownLib")
+  if (ezIsSpecified(linkHtmlLibDir)){
+    file.copy(from=list.files("rmarkdownLib", full.names = TRUE),
+              to="/srv/GT/reference/rmarkdownLib",
+              recursive=TRUE, overwrite=FALSE)
+    unlink("rmarkdownLib", recursive = TRUE)
+    file.symlink(from = linkHtmlLibDir,
+                 to = "rmarkdownLib")
+  }
 }
